feat: make train_rnn_crf work with acc/f1 and weights
